Fixed issue that broke test
The Rabi anaysis was assuming that the guess function returns a list for the frequency guess.
This is wrong, since the guess function itself should (and now does) take care of picking the
correct guess if there are several possibilities.
Also, I removed negative frequencies from the spectrum used to guess the frequency.