New Martinis pulse update
lambda_2 and lambda_3 are now separate keyword arguments for the new
martinis pulse, since we won't be using more coefficients.
Added clipping to theta waveform if it goes out of [theta_i, 180 deg],
since that would produce poles in the frequency or voltage waveform.
Added test for new martinis pulse.