fix ordering of channels in UHFQC detector function (#355)
the channels where put in an array by looping over dict.keys(), which was not guaranteed to preserve ordering.

It should work. The ordering is determined in a bit of a round-about way, and could and should probably be arranged a bit clearer. This PR is just a minimum-number-of-lines hotfix of the problem.

Here is how it goes:
UHFQC.acquisition_initialize sets UHFQC.acquisition_paths (strings) according to a list of channels given to it (which are integers), (preserving the order).

acquisition_poll returns a dictionary; the keys of that dictionary are integers, but they are not channels, but indices into acquisition_paths (and thus channels, which has the same order). Looping over these keys as sorted when interpreting the data in the detector thus arranges the returned data in the same order in which channels are passed to acquisition_initialize.