Redesigned how the CZ sim gets the anharmonicity of the static qubit
This breaks previous code but it gets the value from where it should be
stored: the fluxlutman of the static qubit.
It will also be useful if some other parameters from the fluxlutman of
the static qubit are needed in the future.

'w_q1_sweetspot' and 'w_q1_sweetspot' (needed for noise sim) could also
be retrieved in a similar way from the fluxlutman of the static qubit.