WARNING: this might break interactions with physical instruments
This makes possible to optimize over integer parameters and at the same
time keep the Integer validators for parameters to work.

It is a complicated issue to have mixed types in numpy array, validators
to work and data being properly saved in the HDF5 files.

For now the only minor issue that this might have is that the integers
array are still saved as float64 in the HDF5 file. Should be ok and if
it rises problems when retrieving the values back the user should be
able to notice or get the validators to complain.